Update billing page
